🧱 refactor: Rename Build Info Module
